requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgi-redis==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
autobahn==20.6.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.7
django-braces==1.14.0
django-channels==0.7.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-filter==2.3.0
django-heroku==0.3.1
django-messages==0.6.0
django-model-utils==4.0.0
django-online-users==0.3
django-private-chat==0.3.0
django-redis==4.12.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
hiredis==1.0.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack-python==0.5.6
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
redis==2.10.6
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
twilio==6.44.0
txaio==20.4.1
urllib3==1.25.9
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.1.0
